FIX: Refactor code out of model.rs
