+++ base/src/lib.rs
@@ -46,6 +46,7 @@ mod arithmetic;
 mod cast;
 mod conditional_formatting;
 mod constants;
+mod cut_paste;
 mod functions;
 mod implicit_intersection;
 mod model;
